Improved on Room descriptions, added new rooms, added scenery and Items
In this session I worked hard on the descriptions of the rooms, as well
as started to build some of the scenery. For most of the rooms I was
able to build the scenery and items, but have yet to finish their
descriptions. I also tried to work on making the NPC characters as well
as adding "instead of commands" but those didn't work correctly, so I
commented them out, and will ask for help in class.