Worked on NPCs, creating talking action, more descriptions/backstory, and clothing.
This time in my project I developed my backstory and added a ton to
descriptions of the things in the rooms. I also added secrets that
tells of the history of the character. I was trying to work on my NPC
but that was working, so I went with the suggestion to leave them
non-talkable. I also started to develope my missions and will continue
to do so next time.