Worked on fixing directions of rooms, pulling levers/moving character, and improving descriptions
This time I worked on my rooms because when I tested my game it was
skipping over a room because the directions were messed up. So, this
time I worked on fixing that because it was skipping over a key part. I
also improved on pulling the lever to go to the moon, but it wasn't
working, so I will get help in class. I also worked on improving
descriptions.