Got Score Bar, Action to Unlock door, moving player, winning game, and instead of commands that need Items.
This time I worked on adding an extra that created the score bar. I
found out how to do this in the inform handbook. I also got help in
class on how to unlock the partition between rooms by getting in the
centrifuge. Before I had made pulling a lever to launch the rocket an
out of world action, but Mrs. Kiang helped me to change it because it
did not need to be out of world. Tyler helped me with the digging part,
because I remembered that in his mini-game he had digging sand and I
wanted to implement that command with the moon. This action he taught
me was that you needed to make whatever you were digging a closed
openable container. I also made this how to win.