Worked on scenery and things on moons surface, creating things to be used later on, talking to NPC, more backstory, and fixing errors.
This time I worked more on the scenery of the second part of the game
where you land on the moon. These things also contributed to the player
finding out that he is a russian. I also started adding more things in
the game that will be used for point, and tried to get my skeleton key
to work, but it wouldn't allow me to get the things in the container so
I just made it for one container. I also worked on greg the NPC who
gives you missions.