got else commands working, almost finished missions, tried to make connecting action, and winning.
This time I got my else commands working because they weren't before in
two places, but I copied them from another part of my game and just
replace what it said and it worked. I also almost finished my missions,
and since I couldn't figure out how to make the player to have more
than one item that they needed for an instead of command, I split it
into two missions. I also tried to make one of my missions with the
action of connecting, but it didn't work. I will try to get help in
class. I also got the winning to work because before I had forgotten to
put in a location so it understood the command but nothing would happen
because of it.